+++ resources/js/pages/admin/role/Role.vue
@@ -3,7 +3,7 @@
       <!-- START PAGE CONTENT-->
       <div class="page-heading row">
           <breadcrumb :title='title' class="col-6"></breadcrumb>
-          <router-link :to="{name:'add.role'}" class="col-6 text-right mt-5"><button type="button" class="btn btn-primary">Add New</button></router-link>
+          <router-link v-if="$can('create roles')" :to="{name:'add.role'}" class="col-6 text-right mt-5"><button type="button" class="btn btn-primary">Add New</button></router-link>
       </div>
       <div class="page-content fade-in-up">
         <div class="ibox">
@@ -35,7 +35,7 @@ export default {
           search: '',
           length: 10,
           column: 'id',
-          dir: 'desc'
+          dir: 'asc'
       },
       columns: [
           {
@@ -96,14 +96,20 @@ export default {
           cancelButtonColor: '#d33',
           confirmButtonText: 'Yes, delete it!'
         }).then((result) => {
-
-          if (result.value) {
-            //Send Request to server
-            this.$store.dispatch('role/deleteRole', id).then(
-                    this.getData(route("roles"), this.tableProps)
-                )
+            if(Permissions.indexOf('delete roles') == -1){
+                Swal.fire({
+                    icon: 'error',
+                    title: 'Oops...',
+                    text: 'Bạn không có quyền xoá vai trò!',
+                })
+            }else{
+                if (result.value) {
+                    //Send Request to server
+                    this.$store.dispatch('role/deleteRole', id).then(
+                            this.getData(route("roles"), this.tableProps)
+                        )
+                }
             }
-
         })
       }
     }
